### 2018-04-25 #700 @whedon
[PRE REVIEW]: EndoMineR for the extraction of endoscopic and associated pathology data from medical reports
**Submitting author:** @sebastiz (<a href="http://orcid.org/0000-0003-1673-2663">Sebastian Zeki</a>)
**Repository:** <a href="https://github.com/ropensci/EndoMineR" target ="_blank">https://github.com/ropensci/EndoMineR</a>
**Version:** v1.0.0
