### 2018-04-25 #701 @arfon
Accepting because of https://github.com/ropensci/onboarding/issues/153
