### 2018-04-25 #701 @arfon
@sebastiz - a couple of things:

1. Could you please take a look at the PDF compilation errors above? If you push changes to https://github.com/ropensci/EndoMineR then you can ask Whedon to try and compile the paper with `@whedon generate pdf`
2. Once the paper compiles properly, please make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive?
