### 2018-04-25 #700 @karthik
@sebastiz No additional review required at JOSS. But please check the issues that @arfon has poined out in https://github.com/openjournals/joss-reviews/issues/701
