### 2018-04-25 a06ec76
tabbed bibliography to compile pdf
