+++ paper/paper.bib
@@ -1,21 +1,20 @@
-@article{
-  author = {Bretthauer M. ,Aabakken L. ,Dekker E.,et al.} 
-  title = {{Reporting systems in gastrointestinal endoscopy: Requirements and standards facilitating quality improvement: European Society of Gastrointestinal Endoscopy position statement.}},
-  year = {2016},
-  journaltitle = {United European Gastroenterology Journal},
-  howpublished = {\url{https://doi.org/10.6084/m9.figshare.828487}}
-  DOI = {10.1177/2050640616629079}
-}
+  @article{
+    author = {Bretthauer M. ,Aabakken L. ,Dekker E.,et al.} 
+    title = {{Reporting systems in gastrointestinal endoscopy: Requirements and standards facilitating quality improvement: European Society of Gastrointestinal Endoscopy position statement.}},
+    year = {2016},
+    journaltitle = {United European Gastroenterology Journal},
+    howpublished = {\url{https://doi.org/10.6084/m9.figshare.828487}}
+    DOI = {10.1177/2050640616629079}
+  }
 
 
-
-@article{
-  author = {Stebbing JF.},
-  title = {{Quality assurance of endoscopy units. Best Practice and Research}},
-  year = {2011},
-  journaltitle = {Clinical Gastroenterology},
-  howpublished = {\url{https://doi.org/10.1016/j.dld.2012.06.021},
-  doi = {10.1016/j.dld.2012.06.021}
-}
+  @article{
+    author = {Stebbing JF.},
+    title = {{Quality assurance of endoscopy units. Best Practice and Research}},
+    year = {2011},
+    journaltitle = {Clinical Gastroenterology},
+    howpublished = {\url{https://doi.org/10.1016/j.dld.2012.06.021},
+    doi = {10.1016/j.dld.2012.06.021}
+  }
 
 
