### 2018-04-25 c3023ea
Further typos for formatting bibiliography now corrected
