+++ paper/paper.bib
@@ -1,9 +1,9 @@
   @article{
-    author = {Bretthauer M. ,Aabakken L. ,Dekker E.,et al.} 
+    author = {Bretthauer M. ,Aabakken L. ,Dekker E.,et al.},
     title = {{Reporting systems in gastrointestinal endoscopy: Requirements and standards facilitating quality improvement: European Society of Gastrointestinal Endoscopy position statement.}},
     year = {2016},
     journaltitle = {United European Gastroenterology Journal},
-    howpublished = {\url{https://doi.org/10.6084/m9.figshare.828487}}
+    howpublished = {\url{https://doi.org/10.6084/m9.figshare.828487}},
     DOI = {10.1177/2050640616629079}
   }
 
