### 2018-04-26 457a519
Change bib formatting
