+++ paper/paper.bib
@@ -1,5 +1,5 @@
   @article{
-    author = {Bretthauer M. ,Aabakken L. ,Dekker E.,et al.},
+    author = {Bretthauer M. ,Aabakken L. ,Dekker E., et al.},
     title = {{Reporting systems in gastrointestinal endoscopy: Requirements and standards facilitating quality improvement: European Society of Gastrointestinal Endoscopy position statement.}},
     year = {2016},
     journaltitle = {United European Gastroenterology Journal},
