+++ paper/paper.bib
@@ -1,4 +1,4 @@
-  @article{
+  @article{Bretthauer2016Reporting,
     author = {Bretthauer M. ,Aabakken L. ,Dekker E., et al.},
     title = {{Reporting systems in gastrointestinal endoscopy: Requirements and standards facilitating quality improvement: European Society of Gastrointestinal Endoscopy position statement.}},
     year = {2016},
@@ -8,7 +8,7 @@
   }
 
 
-  @article{
+  @article{Stebbing2011quality,
     author = {Stebbing JF.},
     title = {{Quality assurance of endoscopy units. Best Practice and Research}},
     year = {2011},
