### 2018-04-26 9ff945b
Further bib formatting changes
