+++ paper/paper.bib
@@ -13,7 +13,7 @@
     title = {{Quality assurance of endoscopy units. Best Practice and Research}},
     year = {2011},
     journaltitle = {Clinical Gastroenterology},
-    howpublished = {\url{https://doi.org/10.1016/j.dld.2012.06.021},
+    howpublished = {\url{https://doi.org/10.1016/j.dld.2012.06.021}},
     doi = {10.1016/j.dld.2012.06.021}
   }
 
