+++ paper/paper.bib
@@ -9,7 +9,7 @@
 
 
   @article{Stebbing2011quality,
-    author = {Stebbing J.},
+    author = {Stebbing J. ,},
     title = {{Quality assurance of endoscopy units. Best Practice and Research}},
     year = {2011},
     journaltitle = {Clinical Gastroenterology},
