### 2018-04-26 3504f46
References tidied up
