+++ paper/paper.bib
@@ -3,7 +3,6 @@
     title = {{Reporting systems in gastrointestinal endoscopy: Requirements and standards facilitating quality improvement: European Society of Gastrointestinal Endoscopy position statement.}},
     year = {2016},
     journaltitle = {United European Gastroenterology Journal},
-    howpublished = {\url{https://doi.org/10.6084/m9.figshare.828487}},
     DOI = {10.1177/2050640616629079}
   }
 
@@ -13,7 +12,6 @@
     title = {{Quality assurance of endoscopy units. Best Practice and Research}},
     year = {2011},
     journaltitle = {Clinical Gastroenterology},
-    howpublished = {\url{https://doi.org/10.1016/j.dld.2012.06.021}},
     doi = {10.1016/j.dld.2012.06.021}
   }
 
