### 2018-04-27 0fa9ab1
bib reformatting again
