+++ paper/paper.bib
@@ -1,18 +1,22 @@
   @article{Bretthauer2016Reporting,
-    author = {Bretthauer M. ,Aabakken L. ,Dekker E., et al.},
     title = {{Reporting systems in gastrointestinal endoscopy: Requirements and standards facilitating quality improvement: European Society of Gastrointestinal Endoscopy position statement.}},
-    year = {2016},
+    author = {Bretthauer M. ,Aabakken L. ,Dekker E., et al.},
     journaltitle = {United European Gastroenterology Journal},
+    volume={4},
+    pages={172-6},
+    year = {2016},
+    month={Apr},
     DOI = {10.1177/2050640616629079}
   }
 
 
   @article{Stebbing2011quality,
+    title = {{Quality assurance of endoscopy units.}},
     author = {Stebbing J. ,},
-    title = {{Quality assurance of endoscopy units. Best Practice and Research}},
-    year = {2011},
-    journaltitle = {Clinical Gastroenterology},
-    doi = {10.1016/j.dld.2012.06.021}
+    journaltitle = {Best Practice and Research Clinical Gastroenterol},
+    volume={25},
+    pages={361-70},
+    year = {2011},361-70. 
+    month={Jun},
+    doi = {10.1016/j.bpg.2011.05.004}
   }
-
-
