+++ paper/paper.bib
@@ -16,7 +16,7 @@
     journaltitle = {Best Practice and Research Clinical Gastroenterol},
     volume={25},
     pages={361-70},
-    year = {2011},361-70. 
+    year = {2011},
     month={Jun},
     doi = {10.1016/j.bpg.2011.05.004}
   }
