### 2018-04-27 #701 @sebastiz
@arfon

I have now proofed the pdf and uploaded to zenodo.
Two questions: Do I put the DOI on the ReadMe for the package
I uploaded a copy of the software from a forked version in my repository as the software is now under rOpenSci and zenodo wouldnt link to that. Is that the correct way to do it?

The DOI is:

doi.org/10.5281/zenodo.1232059
