### 2018-04-27 #701 @sebastiz
@arfon

Sorry.... My last question is how do I get the paper registered on pubmed?

Thanks
Sebastian Zeki
