### 2018-04-27 #701 @arfon
> Sorry.... My last question is how do I get the paper registered on pubmed?

I'm not sure what the current status of this. I know some of our editors were working on this over in https://github.com/openjournals/joss/issues/153
